url_safety_chatbot/requirements.txt
streamlit==1.37.0
requests>=2.31.0
tldextract>=5.1.2
python-whois>=0.9.4
dnspython>=2.6.1
python-dotenv>=1.0.1
